Fix bug in ENFA input parameter to specifiy component selection by cumulative variation.  Variation input was set as an int between 50-100, but should have been 0.5-1.0 and stored as a double

+++ src/algorithms/enfa/enfa.hh
@@ -441,7 +441,7 @@ protected:
     int _retainComponents; 
 
     /** the percent variation for selection of components to retain */
-    int _retainVariation; 
+    double _retainVariation; 
 
     /** Minumum number of components required for a valid model */
     int minComponentsInt;
